test(features): add functions feature and refactor steps
