+++ tests/features/bitwise.feature
@@ -1,31 +1,31 @@
 Feature: Skilang bitwise operations
 
-  Scenario: AND operation
+  Scenario: AND operator
     Given the expression "6 & 4"
     When the expression is parsed
     Then the result should be the number "4"
 
-  Scenario: OR operation
+  Scenario: OR operator
     Given the expression "2 | 1"
     When the expression is parsed
     Then the result should be the number "3"
 
-  Scenario: XOR operation
+  Scenario: XOR operator
     Given the expression "5 ^ 3"
     When the expression is parsed
     Then the result should be the number "6"
 
-  Scenario: NOT operation
+  Scenario: NOT operator
     Given the expression "~0"
     When the expression is parsed
     Then the result should be the number "-1"
 
-  Scenario: Left shift operation
+  Scenario: Left shift operator
     Given the expression "3 << 2"
     When the expression is parsed
     Then the result should be the number "12"
 
-  Scenario: Right shift operation
+  Scenario: Right shift operator
     Given the expression "8 >> 1"
     When the expression is parsed
     Then the result should be the number "4"
