+++ tests/features/steps/parsing.py
@@ -22,17 +22,4 @@ def step_expression_parsed(context):
     if "exp" not in context:
         raise ValueError("Expression not provided.")
     parsed: Formula = parse(context.exp)
-    context.parsed_exp = context.result = parsed
-
-
-@when("the expression is evaluated")
-def step_expression_evaluated(context):
-    """
-    :type context: behave.runner.Context
-    """
-    if "parsed_exp" not in context:
-        raise ValueError("Parsed expression not provided.")
-    if "assignments" not in context:
-        raise ValueError("Assignments not provided.")
-    parsed: Formula = context.parsed_exp
-    context.result = parsed.evaluate(**context.assignments)
+    context.result = context.parsed_exp = parsed
