+++ tests/features/steps/results.py
@@ -1,15 +1,37 @@
 from behave import use_step_matcher, then
 
+from skilang import Predicate
 
 use_step_matcher("parse")
 
 
 @then('the result should be the number "{expected}"')
-def step_number_result(context, expected: int):
+def step_number_result(context, expected: str):
     """
     :type context: behave.runner.Context
     :param expected: the expected result
     """
-    if "parsed_exp" not in context:
-        raise ValueError("Parsed expression not provided.")
-    assert float(expected) == context.parsed_exp, f"Expected {float(expected)}, but got {context.parsed_exp}"
+    if "result" not in context:
+        raise ValueError("Actual result not provided.")
+    assert float(expected) == context.result, f"Expected {float(expected)}, but got {context.result}"
+
+
+@then('the result should be the string "{expected}"')
+def step_string_result(context, expected: str):
+    """
+    :type context: behave.runner.Context
+    :param expected: the expected result
+    """
+    if "result" not in context:
+        raise ValueError("Actual result not provided.")
+    assert expected == context.result, f"Expected {expected}, but got {context.result}"
+
+
+@then("the result should be a function")
+def step_function_result(context):
+    """
+    :type context: behave.runner.Context
+    """
+    if "result" not in context:
+        raise ValueError("Actual result not provided.")
+    assert isinstance(context.result, Predicate), f"Expected a function, but got {type(context.result)}"
